[mvsUtils] Move TileParams header from depthMap to mvsUtils
