+++ src/aliceVision/depthMap/CMakeLists.txt
@@ -9,7 +9,6 @@ set(depthMap_files_headers
   Sgm.hpp
   SgmDepthList.hpp
   SgmParams.hpp
-  TileParams.hpp
   volumeIO.hpp
 )
